pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-payment-midtrans"
version = "1.0.0"
description = "A comprehensive Django package for Midtrans Core API payment gateway integration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Danang Haris Setiawan"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "Django>=4.2",
    "djangorestframework>=3.14",
    "requests>=2.28",
    "celery>=5.3",
    "django-celery-beat>=2.5",
]

[project.optional-dependencies]
unfold = [
    "django-unfold>=0.40",
]
dev = [
    "pytest",
    "pytest-django",
    "factory-boy",
    "black",
    "ruff",
]

[project.urls]
Homepage = "https://github.com/danangharissetiawan/django-payment-midtrans"
Documentation = "https://github.com/danangharissetiawan/django-payment-midtrans#readme"
Repository = "https://github.com/danangharissetiawan/django-payment-midtrans"

[tool.setuptools.packages.find]
include = ["django_midtrans*"]

[tool.ruff]
line-length = 120
target-version = "py310"
